feat: add entry component added
AddEntry component added that gets metric data from helpers and each object returned from helper object is included in the state with a default value of 0 as starting point. Also, an increment and decrement methods are added inside the component to let the metric value be changed as per the steps for each metric. A slider method is added for sleep and eat metric